Update DESIGN OF EGG HATCHING DETECTION FEATURE USING RTC DS3231 AND RFID TAG TO OPTIMIZE AUTOMATIC INCUBATOR SYSTEM PERFORMANCE
The poultry farming industry in Indonesia is rapidly growing, but egg hatching management is still hindered by time constraints and farmer supervision. Optimal hatching is important for producing quality chicken chicks, but it is often not achieved. An automatic incubator with the ability to regulate humidity and air circulation becomes the right solution. This research designs an IoT-based automatic incubator with the MQTT protocol featuring an RTC DS3231 for time recording, RFID for egg identification, a microphone sensor for detecting the sound of shell cracking, and a DHT11 sensor for monitoring temperature and humidity. This system is integrated with Thinger.io for remote real-time monitoring and control. Test results show that the system is capable of recording data with high accuracy, temperature of 36–38°C, humidity of 55–68%, and detecting crack sounds with a sensitivity of 85–100%. RFID successfully identifies eggs within a distance of 3 cm, and the egg rack rotation mechanism ensures even distribution of temperature and humidity. This system improves efficiency, hatching success, and reduces the risk of human error, supporting the quality and productivity of the poultry farming industry in Indonesia.